+++ crates/surface-engine/src/goals.rs
@@ -47,7 +47,7 @@ use gandr_core_checker::syntax::Term;
 use gandr_core_checker::syntax::Value;
 use gandr_core_checker::types::Ty;
 use gandr_core_checker::types::ValueType;
-use gandr_core_incrementality::region::Item;
+use gandr_core_incremental::region::Item;
 
 use crate::boundary::ContextLength;
 use crate::boundary::ItemIndex;