+++ crates/surface-engine/src/lower/codata.rs
@@ -65,7 +65,7 @@ use gandr_core_checker::syntax::Value;
 use gandr_core_checker::types::CompType;
 use gandr_core_checker::types::Ty;
 use gandr_core_checker::types::ValueType;
-use gandr_core_incrementality::region::Item;
+use gandr_core_incremental::region::Item;
 
 use super::COut;
 use super::EOut;